// src/@types/user.d.ts
export default interface IUser {
  id: string | number;
  name: string;
  businessName: string;
  firstname: string;
  mail: string;
  errorEmail: boolean;
  phone: string;
  errorPhone: boolean;
  password: string;
  confirmPassword: string;
  errorPassword: boolean;
  isConnected: boolean;
  statutProfil: boolean;
  randomKey: string;
  numberStreet: string;
  street: string;
  codeZip: string;
  city: string;
  idImageAvatar: string;
  idImagePermis: string;
  titleImagePermis: string;
  idImageValidation: string;
  titleImageValidation: string;
  idImageAssurance: string;
  titleImageAssurrance: string;
  errorFormatImagePermis: boolean;
  errorFormatImageValidation: boolean;
  errorFormatImageAssurance: boolean;
  isTraqueur: boolean;
  isPoste: boolean;
  isPetitGibier: boolean;
  isPiegeur: boolean;
  isRechercheAuSang: boolean;
  idParcPhotoProfil: string;
  idParcPhotoIllustrationOne: string;
  idParcPhotoIllustrationTwo: string;
  idParcPhotoIllustrationThree: string;
  statusParcPhotoProfil: string;
  statusParcIllustrationOne: string;
  statusParcIllustrationTwo: string;
  statusParcIllustrationThree: string;
  linkWeb: string;
  linkFacebook: string;
  linkInstagram: string;
  textDescription: string;
  textMailDescription: string;
  mailNotification: [];
  mailNotificationIdOne: string;
  openDate: string;
  closeDate: string;
  isMonday: boolean;
  mondayHoursOpen: string;
  mondayHoursClose: string;
  isTuesday: boolean;
  tuesdayHoursOpen: string;
  tuesdayHoursClose: string;
  isWednesday: boolean;
  wednesdayHoursOpen: string;
  wednesdayHoursClose: string;
  isThursday: boolean;
  thursdayHoursOpen: string;
  thursdayHoursClose: string;
  isFriday: boolean;
  fridayHoursOpen: string;
  fridayHoursClose: string;
  isSaturday: boolean;
  saturdayHoursOpen: string;
  saturdayHoursClose: string;
  isSunday: boolean;
  sundayHoursOpen: string;
  sundayHoursClose: string;
  delayRdv: number | null;
  delayAnnulaitionRdv: number | null;
  isSendForNewRdv: boolean;
  textConsigne: string;
  isDisplayConsigne: boolean;
  isSendNotification: boolean;
  listFollows: [];
  listMailInvitations: [];
  listMembers: IUserMember[];
}